// src/types/carrier/transactionInfo.ts
export interface IDetailTransaction {
  id: number;
  trans_type: number;
  carrier_operator_id: string;
  carrier_operator_code: string;
  carrier_operator_name: string;
  shipper_operator_id: string;
  shipper_operator_code: string;
  shipper_operator_name: string;
  carrier2_operator_id: any;
  carrier2_operator_code: any;
  carrier2_operator_name: string;
  req_cns_line_item_id: number;
  cns_line_item_id: any;
  cns_line_item_by_date_id: number;
  vehicle_avb_resource_id: number;
  vehicle_avb_resource_item_id: number;
  vehicle_diagram_item_id: number;
  trailer_number: number;
  departure_from: number;
  arrival_to: number;
  transport_date: string;
  request_price: any;
  request_collection_time_from: any;
  request_collection_time_to: any;
  propose_price: any;
  propose_departure_time: any;
  propose_arrival_time: any;
  shipper_contract_file: any;
  carrier_contract_file: any;
  carrier2_contract_file: any;
  contract_file: any;
  status: number;
  trsp_cli_prty_name_txt: string;
  trans_matching_id: number;
  cnsg_prty_name_txt: any;
  item_name_txt: string;
  road_carr_depa_sped_org_name_txt: string;
  trsp_cli_tel_cmm_cmp_num_txt: string;
  is_request_carrier: boolean;
  sip_track_id: string;
  request_snapshot: {
    id: number;
    weight: number;
    status: number;
    operator_id: string;
    operator_code: string;
    req_cns_line_item_id: number | null;
    trans_plan_id: number | null;
    transport_code: string;
    transport_name: string;
    collection_date: string;
    collection_time_from: string;
    collection_time_to: string;
    departure_from: number;
    arrival_to: number;
    trailer_number: number;
    trailer_number_rest: number;
    price_per_unit: number;
    outer_package_code: number;
    total_length: number;
    total_width: number;
    total_height: number;
    special_instructions: number;
    trans_type: number;
    trans_order_id: number;
    cns_line_item_id: number;
    cut_off_time?: string;
    cut_off_fee?: string;
    temperature_range: [];
  };
  propose_snapshot: {
    id: number;
    day: string;
    price: number;
    status: number;
    giai: string;
    operator_id: string;
    operator_code: string;
    vehicle_avb_resource_id: string | null;
    vehicle_diagram_id: string | null;
    departure_from: number;
    arrival_to: number;
    trip_name: string;
    departure_time_min: string | null;
    departure_time_max: string | null;
    arrival_time: string;
    adjustment_price: number | null;
    vehicle_type: number | null;
    display_order: number | null;
    assign_type: number;
    max_payload: number;
    total_length: number;
    total_width: number;
    total_height: number | null;
    car_max_load_capacity1_meas: string | null;
    vehicle_size: number;
    temperature_range: [];
    cut_off_time?: any;
    cut_off_fee?: any;
  };
  negotiation_data?: {
    shipper: any;
    carrier: any;
  };
  car_info: CarInfo;
  latest_message?: MESSAGE_TRANSACTIONS;
  shipper_info?: IShipperInfo;
  created_date: string;
  total: number;
  parent_order_id: number;
  parent_order: string;
  trip_name: string;
  trailer_license_plt_num_id: string;
  sibling_order_status: number;
  departure_time: string;
  arrival_time: string;
}

export type MESSAGE_TRANSACTIONS = {
  roomId: number;
  company_name: string;
  from: string;
  message: string;
  time: string;
  date: string;
  role: string;
};

export type CarInfo = [
  {
    car_lngh_meas: string;
    car_wid_meas: string;
    car_trk_bed_lngh_meas: string;
    car_max_untl_cp_quan: string;
    car_weig_meas: string;
    car_vol_of_hzd_item_meas: string;
    car_trk_bed_hght_meas: string;
    trms_of_upp_tmp_meas: string;
    car_spc_grv_of_hzd_item_meas: string;
    id: string;
    car_cls_of_rfg_exst_cd: string;
    car_cls_of_shp_cd: string;
    service_strt_date: string;
    pcke_frm_name_cd: string;
    car_trk_bed_grnd_hght_meas: string;
    service_end_date: string;
    tractor_idcr: string;
    service_name: string;
    car_cls_of_tlg_lftr_exst_cd: string;
    car_max_load_capacity1_meas: string;
    car_cls_of_crn_exst_cd: string;
    pcke_frm_cd: string;
    trms_of_lwr_tmp_meas: string;
    car_cmpn_name_of_gtp_crtf_exst_txt: string;
    freight_rate: string;
    car_trk_bed_wid_meas: string;
    car_hght_meas: string;
    car_license_plt_num_id: any;
    car_body_num_cd: string;
    car_max_load_capacity2_meas: string;
    car_max_load_vol_meas: string;
    car_rmk_about_eqpm_txt: string;
    service_strt_time: string;
    service_no: string;
    giai: string;
    car_cls_of_size_cd: string;
    service_end_time: string;
    car_cls_of_wing_body_exst_cd: string;
    car_ctrl_num_id: string;
    trailer_license_plt_num_id: string;
    is_matched: boolean;
    temperature_range: string[] | null;
    vehicle_name: string;
  },
];

export interface IShipperInfo {
  id: number;
  trans_type: number;
  carrier_operator_id: string;
  carrier_operator_code: string;
  carrier_operator_name: string;
  shipper_operator_id: string;
  shipper_operator_code: string;
  shipper_operator_name: string;
  carrier2_operator_id: any;
  carrier2_operator_code: any;
  req_cns_line_item_id: number;
  cns_line_item_id: any;
  cns_line_item_by_date_id: number;
  vehicle_avb_resource_id: number;
  vehicle_avb_resource_item_id: number;
  trailer_number: number;
  departure_from: number;
  arrival_to: number;
  transport_date: string;
  request_price: any;
  request_collection_time_from: any;
  request_collection_time_to: any;
  propose_price: any;
  propose_departure_time: any;
  propose_arrival_time: any;
  shipper_contract_file: any;
  carrier_contract_file: any;
  carrier2_contract_file: any;
  contract_file: any;
  status: number;
  trsp_cli_prty_name_txt: string;
  cnsg_prty_name_txt: any;
  item_name_txt: string;
  is_request_carrier: boolean;
  request_snapshot: {
    id: number;
    weight: number;
    status: number;
    operator_id: string;
    operator_code: string;
    req_cns_line_item_id: number | null;
    transport_plan_id: number | null;
    transport_code: string;
    transport_name: string;
    collection_date: string;
    collection_time_from: string;
    collection_time_to: string;
    departure_from: number;
    arrival_to: number;
    trailer_number: number;
    trailer_number_rest: number;
    price_per_unit: number;
    vehicle_condition: string | null;
    outer_package_code: number;
    total_length: number;
    total_width: number;
    total_height: number;
    special_instructions: number;
    trans_type: number;
    trans_order_id: number;
    cns_line_item_id: number;
    cut_off_time?: string;
    cut_off_fee?: string;
  };
  propose_snapshot: {
    id: number;
    day: string;
    price: number;
    status: number;
    giai: string;
    operator_id: string;
    operator_code: string;
    vehicle_avb_resource_id: string | null;
    vehicle_diagram_id: string | null;
    departure_from: number;
    arrival_to: number;
    trip_name: string;
    departure_time_min: string | null;
    departure_time_max: string | null;
    arrival_time: string;
    adjustment_price: number | null;
    vehicle_type: number | null;
    display_order: number | null;
    assign_type: number;
    max_payload: number;
    total_length: number;
    total_width: number;
    total_height: number | null;
    car_max_load_capacity1_meas: string | null;
    vehicle_size: number;
    temperature_range: number[] | null;
    cut_off_time?: any;
    cut_off_fee?: any;
  };
  car_info: CarInfo;
  latest_message?: MESSAGE_TRANSACTIONS;
}

export interface subTrailerType {
  carrier2_operator_name: string;
  carrier_operator_name: string;
  shipper_operator_name: string;
  isFirstTrailer: boolean;
}